+++ src/ts/infrastructure/ui/index.ts
@@ -6,7 +6,7 @@ export { default as setLoading } from './setLoading';
 export { default as setOffline } from './setOffline';
 export { default as setVersionOutdated } from './setVersionOutdated';
 export { default as updateTabVisibility, hideAllFeatureTabs } from './updateTabVisibility';
-export { default as initializeColorModeToggler } from './DBColorToggler';
+export * from './useColorMode';
 export * from './loadingButtonState';
 export * from './CustomSnackbar';
 export * from './dbButton';